requirements.txt
spotipy
pandas
requests
google-genai
streamlit
python-dotenv
